+++ API/Modulo de ventas y cuenta corriente/Modulo de ventas y cuenta corriente/Models/DETALLES_DE_PAGO.cs
@@ -24,9 +24,9 @@ namespace Modulo_de_ventas_y_cuenta_corriente.Models
         [ForeignKey("ID_BANCO")]
         public int ID_BANCO { get; set; }
 
-        public int MONTO { get; set; }
+        public decimal MONTO { get; set; }
 
-        public int VUELTO { get; set; }
+        public decimal VUELTO { get; set; }
 
         public string NUMERO { get; set; }
