+++ API/Modulo de ventas y cuenta corriente/Modulo de ventas y cuenta corriente/Models/NOTAS_DE_CREDITO.cs
@@ -24,9 +24,9 @@ namespace Modulo_de_ventas_y_cuenta_corriente.Models
 
         public string RAZON { get; set; }
 
-        public int MONTO { get; set; }
+        public decimal MONTO { get; set; }
 
-        public int IVA { get; set; }
+        public decimal IVA { get; set; }
 
     }
 }
