+++ API/Modulo de ventas y cuenta corriente/Modulo de ventas y cuenta corriente/Models/PAGOS.cs
@@ -24,7 +24,7 @@ namespace Modulo_de_ventas_y_cuenta_corriente.Models
 
         public DateTime FECHA { get; set; }
 
-        public int TOTAL { get; set; }
+        public decimal TOTAL { get; set; }
 
         ICollection<DETALLES_DE_PAGO> _DETALLES_DE_PAGO { get; set; }
     }
